Set up enviornment variables for db connection in production - requires refactoring file structure
