+++ .sequelizerc
@@ -1,0 +1,8 @@
+const path = require('path');
+
+module.exports = {
+  config: path.resolve(__dirname, 'db/config/config.js'),
+  'models-path': path.resolve(__dirname, 'db/models'),
+  'seeders-path': path.resolve(__dirname, 'db/seeders'),
+  'migrations-path': path.resolve(__dirname, 'db/migrations')
+};
