--- config/config.json
@@ -1,21 +1,0 @@
-{
-  "development": {
-    "username": "postgres",
-    "database": "tweet_311_service_dev",
-    "host": "127.0.0.1",
-    "dialect": "postgres"
-  },
-  "test": {
-    "username": "postgres",
-    "database": "tweet_311_service_test",
-    "host": "127.0.0.1",
-    "dialect": "postgres"
-  },
-  "production": {
-    "username": "root",
-    "password": null,
-    "database": "tweet_311_service_prod",
-    "host": "127.0.0.1",
-    "dialect": "postgres"
-  }
-}
