+++ lib/api/v1/locations.js
@@ -1,4 +1,4 @@
-var models = require('../../../models');
+var models = require('../../../db/models');
 
 exports.methods = {
   all: async function(request, reply) {
